Removing some warnings on Windows.

+++ reg-lib_gpu/_reg_mutualinformation_gpu.cu
@@ -31,7 +31,7 @@ void reg_getVoxelBasedNMIGradientUsingPW_gpu(   nifti_image *targetImage,
 	const int voxelNumber = targetImage->nvox;
 	const int binNumber = binning*(binning+2);
     const float4 entropies_h=make_float4((float)entropies[0],(float)entropies[1],(float)entropies[2],(float)entropies[3]);
-	const float NMI = (entropies[0]+entropies[1])/entropies[2];
+	const float NMI = (float)((entropies[0]+entropies[1])/entropies[2]);
 
 	// Bind Symbols
 	CUDA_SAFE_CALL(cudaMemcpyToSymbol(c_VoxelNumber,&voxelNumber,sizeof(int)));
